Yet another compile fix for MinGW/Windows.
Reviewed-on: http://codereview.qt.nokia.com/2777

+++ src/libs/utils/qtcprocess.cpp
@@ -690,7 +690,7 @@ BOOL CALLBACK sendShutDownMessageToAllWindowsOfProcess_enumWnd(HWND hwnd, LPARAM
     static UINT uiShutDownMessage = RegisterWindowMessage(L"qtcbuildhelper_shutdown");
     DWORD dwProcessID;
     GetWindowThreadProcessId(hwnd, &dwProcessID);
-    if (lParam == dwProcessID) {
+    if ((DWORD)lParam == dwProcessID) {
         SendNotifyMessage(hwnd, uiShutDownMessage, 0, 0);
         return FALSE;
     }
